+++ CHANGELOG.md
@@ -3,6 +3,14 @@
 This project uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 The format of the release notes follows [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [UNRELEASED]
+
+### Changed
+
+- Improved Logging for `sftp_download` ([#35])
+- Don't consider folder sizes for database size, this is too buggy and inaccurate
+- Use thin spaces between size and unit to log to console and file
+
 ## [v0.6.0] - 2025-12-27
 
 ### Added
@@ -199,3 +207,4 @@ prevent LibCurl error 9
 [#24]: https://github.com/LIM-AeroCloud/ICARE.jl/issues/24
 [#26]: https://github.com/LIM-AeroCloud/ICARE.jl/issues/26
 [#32]: https://github.com/LIM-AeroCloud/ICARE.jl/issues/32
+[#35]: https://github.com/LIM-AeroCloud/ICARE.jl/issues/35